Split history command into its own module
