+++ unison-cli/src/Unison/Codebase/Editor/HandleInput.hs
@@ -65,6 +65,7 @@ import Unison.Codebase.Editor.HandleInput.EditNamespace (handleEditNamespace)
 import Unison.Codebase.Editor.HandleInput.FindAndReplace (handleStructuredFindI, handleStructuredFindReplaceI, handleTextFindI)
 import Unison.Codebase.Editor.HandleInput.FormatFile qualified as Format
 import Unison.Codebase.Editor.HandleInput.Global qualified as Global
+import Unison.Codebase.Editor.HandleInput.History (handleHistory)
 import Unison.Codebase.Editor.HandleInput.InstallLib (handleInstallLib, handleInstallLocalLib)
 import Unison.Codebase.Editor.HandleInput.LSPDebug qualified as LSPDebug
 import Unison.Codebase.Editor.HandleInput.Load (EvalMode (Sandboxed), evalUnisonFile, handleLoad, loadUnisonFile)
@@ -298,29 +299,7 @@ loop e = do
           success <- Cli.popd
           when (not success) (Cli.respond StartOfCurrentPathHistory)
         HistoryI resultsCap diffCap from -> do
-          branch <-
-            case from of
-              BranchAtSCH hash -> Cli.resolveShortCausalHash hash
-              BranchAtPath path' -> do
-                pp <- Cli.resolvePath' path'
-                Cli.getBranchFromProjectPath pp
-              BranchAtProjectPath pp -> Cli.getBranchFromProjectPath pp
-          schLength <- Cli.runTransaction Codebase.branchHashLength
-          history <- liftIO (doHistory schLength 0 branch [])
-          Cli.respondNumbered history
-          where
-            doHistory :: Int -> Int -> Branch IO -> [(CausalHash, Names.Diff)] -> IO NumberedOutput
-            doHistory schLength !n b acc =
-              if maybe False (n >=) resultsCap
-                then pure (History diffCap schLength acc (PageEnd (Branch.headHash b) n))
-                else case Branch._history b of
-                  Causal.One {} -> pure (History diffCap schLength acc (EndOfLog $ Branch.headHash b))
-                  Causal.Merge _ _ _ tails ->
-                    pure (History diffCap schLength acc (MergeTail (Branch.headHash b) $ Map.keys tails))
-                  Causal.Cons _ _ _ tail -> do
-                    b' <- fmap Branch.Branch $ snd tail
-                    let elem = (Branch.headHash b, Branch.namesDiff b' b)
-                    doHistory schLength (n + 1) b' (elem : acc)
+          handleHistory resultsCap diffCap from
         AnnotateI toAnnotate mayMsg -> do
           handleAnnotate toAnnotate mayMsg
         UndoI -> do
